requirements.txt
pandas==1.5.3
SQLAlchemy==1.4.46   # Any >=1.4.28, <2.0 works
psycopg2-binary==2.9.9
apache-airflow==2.7.2
apache-airflow-providers-postgres==5.2.0
